test(config/appdir): partition the tests across platforms
The test suite was not ready for a Windows build: unix-only helpers
sat in a shared source, and neutral rows hard-coded the path
separator. Address both ahead of the Windows implementation.

appdir_test.go asserts FHS and XDG paths, and removes the working
directory to provoke a filepath.Abs failure — none of which holds on
Windows, so it already carries a !windows tag. The XDG environment
helpers it relies on move out of the shared testutils_test.go into a
new testutils_unix_test.go behind that tag; appdir_test.go itself
only gains a note on why its working-directory case cannot run on
Windows.

Join and joinFn test expectations are declared slash-style and
normalised through filepath.FromSlash, so the rows hold whatever the
path separator is and run unchanged everywhere.

No behaviour changes: this partitions the unix-only helpers and
normalises the neutral rows, leaving the tests ready for the Windows
implementation to follow.
